Adds auto-scaling policies
Basic CPU limits are used for scale up/down - we may tune
this based on further tests. SW deployments
which trigger ansible and DNS hostname setting are
now inside the autoscaling group - the reason is that
autoscaling UPDATE events are not propagated to upper
stacks.

A negative side effect of moving SW deployments into
autoscaling group is that these are now called per each node
being added/removed - to minimize number of re-runs of
ansible script it now checks if some changes have been really done
in the inventory file.

Autoscaling is desibled by default and can be enabled by
setting autoscaling=true param.

+++ fragments/master-ansible.sh
@@ -5,6 +5,8 @@ set -eu
 set -x
 set -o pipefail
 
+echo $NODE_HOSTNAME >> /var/lib/openshift_nodes
+
 export HOME=/root
 cat << EOF > /var/lib/ansible-inventory
 # Create an OSEv3 group that contains the masters and nodes groups
@@ -48,17 +50,35 @@ $MASTER_HOSTNAME.$DOMAINNAME openshift_hostname=$MASTER_HOSTNAME.$DOMAINNAME ope
 $MASTER_HOSTNAME.$DOMAINNAME openshift_hostname=$MASTER_HOSTNAME.$DOMAINNAME openshift_public_hostname=$MASTER_HOSTNAME.$DOMAINNAME openshift_node_labels="{'region': 'infra', 'zone': 'default'}"
 EOF
 
+# this script is triggered for each node being added, let's
+# give all nodes some time to write their hostnames into the list (this
+# minimizes number of ansible-playbook re-runs)
+sleep 60
+
 # Write each node
-for node in $NODE_HOSTNAMES
-do
+for node in `cat /var/lib/openshift_nodes`;do
   #echo "$node" >> /var/lib/ansible-inventory
   echo "$node openshift_hostname=$node openshift_public_hostname=$node openshift_node_labels=\"{'region': 'primary', 'zone': 'default'}\"" >> /var/lib/ansible-inventory
 done
 
+while pidof -x /bin/ansible-playbook; do
+  echo "waiting for another ansible-playbook to finish"
+  sleep 10
+done
+
+if [ -e /var/lib/ansible-inventory.deployed ] && diff /var/lib/ansible-inventory /var/lib/ansible-inventory.deployed; then
+    echo "inventory file has not changed since last ansible run, no need to re-run"
+    exit 0
+fi
+
+cp /var/lib/ansible-inventory /var/lib/ansible-inventory.started
+
 # NOTE: docker-storage-setup hangs during cloud-init because systemd file is set
 # to run after cloud-final.  Temporarily move out of the way (as we've already done storage setup
-mv /usr/lib/systemd/system/docker-storage-setup.service $HOME
-systemctl daemon-reload
+if [ -e /usr/lib/systemd/system/docker-storage-setup.service ]; then
+    mv /usr/lib/systemd/system/docker-storage-setup.service $HOME
+    systemctl daemon-reload
+fi
 
 # NOTE: Ignore the known_hosts check/propmt for now:
 export ANSIBLE_HOST_KEY_CHECKING=False
@@ -67,3 +87,5 @@ ansible-playbook --inventory /var/lib/ansible-inventory $HOME/openshift-ansible/
 # Move docker-storage-setup unit file back in place
 mv $HOME/docker-storage-setup.service /usr/lib/systemd/system
 systemctl daemon-reload
+
+mv /var/lib/ansible-inventory.started /var/lib/ansible-inventory.deployed
